1-sem/RAS/Trabalho/Parte3/PictuRAS-main/backend/tools/watermark/requirements.txt
pika==1.3.2
Pillow==11.1.0
python_magic==0.4.27
